rename AliasToClassRector to EntityAliasToClassConstantReferenceRector
