--- packages/Doctrine/tests/Rector/AliasToClassRector/Source/DoctrineEntityManager.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Doctrine\Tests\Rector\AliasToClassRector\Source;
-
-final class DoctrineEntityManager
-{
-
-}
