+++ packages/Doctrine/tests/Rector/MethodCall/EntityAliasToClassConstantReferenceRector/Source/DoctrineEntityManager.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Doctrine\Tests\Rector\MethodCall\EntityAliasToClassConstantReferenceRector\Source;
+
+final class DoctrineEntityManager
+{
+
+}
